requirements.txt
rsa
xmpppy
numpy
python-dotenv
fastapi
mysql-connector-python
paramiko
slixmpp
asyncio